session_catalog: carve caches.rs (pure move)
Filesystem fingerprints and the session-list cache tier (row, codex,
baseline, and intendant caches; persisted session index + preload) —
with tests, verbatim. Purity-checked; tests conserved at 104.
